SDK-1300: Move errors to docscan/response/docscanerr, to prevent circular import when used from other docscan subpackages
