+++ docscan/sandbox/client.go
@@ -7,7 +7,7 @@ import (
 	"os"
 
 	"github.com/getyoti/yoti-go-sdk/v3/cryptoutil"
-	"github.com/getyoti/yoti-go-sdk/v3/docscan"
+	"github.com/getyoti/yoti-go-sdk/v3/docscan/response/docscanerr"
 	"github.com/getyoti/yoti-go-sdk/v3/docscan/sandbox/request"
 	"github.com/getyoti/yoti-go-sdk/v3/requests"
 	yotirequest "github.com/getyoti/yoti-go-sdk/v3/requests"
@@ -79,7 +79,7 @@ func (client *Client) makeConfigureResponseRequest(request *http.Request) error 
 	response, err := requests.Execute(client.getHTTPClient(), request)
 
 	if err != nil {
-		return docscan.NewError(err, response)
+		return docscanerr.New(err, response)
 	}
 
 	return nil
